requirements.txt
streamlit==1.36.0
pydantic==2.7.4
python-docx==1.1.2
pdfminer.six==20231228
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
rapidfuzz==3.8.1
nltk==3.8.1